backend/requirements.txt
flask
flask-limiter
python-dotenv
requests
cryptography
flask-cors
supabase
gunicorn
pytest
stripe
